moved plant sorting logic from harvester.js to field.js
