+++ src/scripts/field.js
@@ -13,3 +13,21 @@ export const addPlant = seed => {
 export const usePlants = () => {
   return plants.slice();
 };
+
+export const usePlantsSorted = () => {
+  return sortPlants(plants);
+};
+
+/**
+ * Given an array of plant objects, returns a new array of plants sorted alphabetically by type, case-insensitive.
+ * @param {Array} plants Array of plant objects to sort 
+ */
+const sortPlants = plants => {
+  const sortedPlants = plants.slice();
+
+  sortedPlants.sort((plantA, plantB) => 
+    plantA.type.toLowerCase().localeCompare(plantB.type.toLowerCase())
+  );
+
+  return sortedPlants;
+};
