+++ src/scripts/harvester.js
@@ -11,19 +11,5 @@ export const harvestPlants = plants => {
     }
   });
 
-  return sortPlants(seeds);
-};
-
-/**
- * Given an array of plant objects, returns a new array of plants sorted alphabetically by type, case-insensitive.
- * @param {Array} plants Array of plant objects to sort 
- */
-const sortPlants = plants => {
-  const sortedPlants = plants.slice();
-
-  sortedPlants.sort((plantA, plantB) => 
-    plantA.type.toLowerCase().localeCompare(plantB.type.toLowerCase())
-  );
-
-  return sortedPlants;
+  return seeds;
 };
