+++ src/scripts/main.js
@@ -1,6 +1,6 @@
 import { createPlan } from './plan.js';
 import { plantSeeds } from './tractor.js';
-import { usePlants } from './field.js';
+import { usePlantsSorted } from './field.js';
 import { harvestPlants } from './harvester.js';
 import { Catalog } from './catalog/Catalog.js';
 
@@ -8,7 +8,7 @@ const yearlyPlan = createPlan();
 
 plantSeeds(yearlyPlan);
 
-const plants = usePlants();
+const plants = usePlantsSorted();
 const harvest = harvestPlants(plants);
 
 Catalog(harvest);
